test(cli): split connect recovery suite
